Move network module tests into own dir
Delete redundant tests
