+++ tests/integration/network/test_resolvers.py
